Desarrollo de VistaCliente (interfaz y logica),
Retocar CrearCliente.kt, CrearTarea diseño, AdaptadorClientes.kt para obtener un cliente para la tarea

Añadir ListaClientes.kt

+++ app/src/main/java/com/example/infotasks/Utiles/ConversorQueryAModelo.kt
@@ -77,11 +77,9 @@ object ConversorQueryAModelo {
             dato!!["dni"].toString(),
             dato["nombre"].toString(),
             dato["apellidos"].toString(),
-            dato["telefono"].toString(),
+            dato["telefono"].toString().toInt(),
             dato["localidad"].toString(),
             dato["domicilio"].toString(),
-
-
         )
     }
